Syncing tree state when visiting a file opens a folder.
This is done by simulating a client request of opening a folder. To reuse
existing code, an extra parameter is added to folderOpened in tree.js to
indicate whether or not to notify the original connection of folder contents.

Add another parameter to sendUserExtendedMessage to allow sending to all
but one connection.